+++ internal/integration/scripts_test.go
@@ -47,9 +47,13 @@ func TestScripts(t *testing.T) {
 		{"v0.3.0/fibonacci.gt", objects.Int(144)},
 		{"v0.3.0/gcd.gt", objects.Int(1)},
 		{"v0.3.0/showcase.gt", objects.Bool(true)},
+
 		{"v0.4.0/function_scale.gt", objects.Int(5136)},
 		{"v0.4.0/fibonacci_closure.gt", objects.Int(34)},
 		{"v0.4.0/polynomials.gt", objects.Bool(true)},
+
+		{"v0.5.0/collatz.gt", objects.Int(104)},
+		{"v0.5.0/fibonacci_closure_loop.gt", objects.Int(34)},
 	}
 
 	for _, tt := range tests {