+++ internal/token/tokentype_string.go
@@ -22,31 +22,32 @@ func _() {
 	_ = x[Bool-11]
 	_ = x[Return-12]
 	_ = x[Print-13]
-	_ = x[Plus-14]
-	_ = x[Minus-15]
-	_ = x[Star-16]
-	_ = x[Slash-17]
-	_ = x[Equal-18]
-	_ = x[Less-19]
-	_ = x[LessEqual-20]
-	_ = x[Greater-21]
-	_ = x[GreaterEqual-22]
-	_ = x[EqualEqual-23]
-	_ = x[NotEqual-24]
-	_ = x[And-25]
-	_ = x[Or-26]
-	_ = x[Not-27]
-	_ = x[SemiColon-28]
-	_ = x[LeftParen-29]
-	_ = x[RightParen-30]
-	_ = x[LeftCurlyBrace-31]
-	_ = x[RightCurlyBrace-32]
-	_ = x[Comma-33]
+	_ = x[While-14]
+	_ = x[Plus-15]
+	_ = x[Minus-16]
+	_ = x[Star-17]
+	_ = x[Slash-18]
+	_ = x[Equal-19]
+	_ = x[Less-20]
+	_ = x[LessEqual-21]
+	_ = x[Greater-22]
+	_ = x[GreaterEqual-23]
+	_ = x[EqualEqual-24]
+	_ = x[NotEqual-25]
+	_ = x[And-26]
+	_ = x[Or-27]
+	_ = x[Not-28]
+	_ = x[SemiColon-29]
+	_ = x[LeftParen-30]
+	_ = x[RightParen-31]
+	_ = x[LeftCurlyBrace-32]
+	_ = x[RightCurlyBrace-33]
+	_ = x[Comma-34]
 }
 
-const _TokenType_name = "InvalidTokenEOFNumberIdentifierLetTrueFalseIfElseFnIntBoolReturnPrintPlusMinusStarSlashEqualLessLessEqualGreaterGreaterEqualEqualEqualNotEqualAndOrNotSemiColonLeftParenRightParenLeftCurlyBraceRightCurlyBraceComma"
+const _TokenType_name = "InvalidTokenEOFNumberIdentifierLetTrueFalseIfElseFnIntBoolReturnPrintWhilePlusMinusStarSlashEqualLessLessEqualGreaterGreaterEqualEqualEqualNotEqualAndOrNotSemiColonLeftParenRightParenLeftCurlyBraceRightCurlyBraceComma"
 
-var _TokenType_index = [...]uint8{0, 12, 15, 21, 31, 34, 38, 43, 45, 49, 51, 54, 58, 64, 69, 73, 78, 82, 87, 92, 96, 105, 112, 124, 134, 142, 145, 147, 150, 159, 168, 178, 192, 207, 212}
+var _TokenType_index = [...]uint8{0, 12, 15, 21, 31, 34, 38, 43, 45, 49, 51, 54, 58, 64, 69, 74, 78, 83, 87, 92, 97, 101, 110, 117, 129, 139, 147, 150, 152, 155, 164, 173, 183, 197, 212, 217}
 
 func (i TokenType) String() string {
 	idx := int(i) - 0